put removePendingFields... back whoops
