+++ src/libs/actions/Policy/DistanceRate.ts
@@ -20,7 +20,7 @@ import type {Policy, Report} from '@src/types/onyx';
 import type {ErrorFields} from '@src/types/onyx/OnyxCommon';
 import type {Attributes, CustomUnit, Rate} from '@src/types/onyx/Policy';
 import type {OnyxData} from '@src/types/onyx/Request';
-import {navigateWhenEnableFeature} from './Policy';
+import {navigateWhenEnableFeature, removePendingFieldsFromCustomUnit} from './Policy';
 
 type NewCustomUnit = {
     customUnitID: string;
@@ -61,18 +61,6 @@ Onyx.connect({
         allPolicies[key] = val;
     },
 });
-
-/**
- * Takes removes pendingFields and errorFields from a customUnit
- */
-function removePendingFieldsFromCustomUnit(customUnit: CustomUnit): CustomUnit {
-    const cleanedCustomUnit = {...customUnit};
-
-    delete cleanedCustomUnit.pendingFields;
-    delete cleanedCustomUnit.errorFields;
-
-    return cleanedCustomUnit;
-}
 
 /**
  * Takes array of customUnitRates and removes pendingFields and errorFields from each rate - we don't want to send those via API
